requirements.txt
annotated-types==0.6.0
argcomplete==1.10.3
asttokens==2.4.1
beautifulsoup4==4.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
compressed_rtf==1.0.6
confection==0.1.4
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
docx==0.2.4
docx2txt==0.8
ebcdic==1.1.1
executing==2.0.1
extract-msg==0.28.7
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.21.3
idna==3.6
IMAPClient==2.1.0
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
langcodes==3.3.0
libmagic==1.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
olefile==0.47
packaging==23.2
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20191110
pillow==10.2.0
platformdirs==4.2.0
poppler-utils==0.1.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pytesseract==0.3.10
python-dateutil==2.9.0
python-docx==1.1.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-pptx==0.6.23
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.12.0
smart-open==6.4.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
SpeechRecognition==3.8.1
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
textract==1.6.5
thinc==8.2.3
tokenizers==0.15.2
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
xlrd==1.2.0
XlsxWriter==3.2.0
